Offers action status as computer value
It occurs, that action status as offer attribute is time sensitive, what means
that we shouldn't store it in database (as it depends on the time, that offer
is requested). Instead of that, we will compute action status value every time
somebody ask for it.

This change also include factories introduction in our backend test suite.
Eventually it will greatly reduce line of code require to set up tests.